fsattendance/backend/requirements.txt
Flask==2.2.2
jellyfish==0.9.0
mysql-connector-python==8.0.30
pipenv==2022.11.5
python-dotenv==0.21.0

